requirements.txt
prometheus_client
requests
PyYAML